+++ src/main/java/com/palantir/gradle/versions/ConsistentVersionsPlugin.java
@@ -47,7 +47,6 @@ public class ConsistentVersionsPlugin implements Plugin<Project> {
         });
         project.getPluginManager().apply(VersionsLockPlugin.class);
         project.getPluginManager().apply(VersionsPropsPlugin.class);
-        project.getPluginManager().apply(GetVersionPlugin.class);
         project.getPluginManager().apply(VersionsPropsIdeaPlugin.class);
         project.getPluginManager().apply(CheckUnusedConstraintsPlugin.class);
 
@@ -56,6 +55,8 @@ public class ConsistentVersionsPlugin implements Plugin<Project> {
         extension.getExternalDependencies().register("gradle-consistent-versions", dep -> dep.atLeastVersion("0.9.0"));
 
         project.allprojects(proj -> {
+            proj.getPluginManager().apply(GetVersionPlugin.class);
+
             proj.getPluginManager().withPlugin("java", _plugin -> {
                 proj.getPluginManager().apply(FixLegacyJavaConfigurationsPlugin.class);
             });
