+++ src/main/java/com/palantir/gradle/versions/GetVersionPlugin.java
@@ -16,43 +16,146 @@
 
 package com.palantir.gradle.versions;
 
+import static java.util.stream.Collectors.toList;
+
+import com.google.common.base.Preconditions;
+import com.google.common.base.Splitter;
+import com.google.common.collect.Iterables;
 import com.palantir.gradle.versions.ConsistentVersionsPlugin.GcvAttributes;
+import groovy.lang.Closure;
+import java.util.List;
+import java.util.Map;
+import java.util.Optional;
+import java.util.stream.Collectors;
 import javax.inject.Inject;
 import org.gradle.api.GradleException;
+import org.gradle.api.NamedDomainObjectProvider;
 import org.gradle.api.Plugin;
 import org.gradle.api.Project;
+import org.gradle.api.artifacts.Configuration;
 import org.gradle.api.artifacts.ConfigurationContainer;
+import org.gradle.api.artifacts.ModuleVersionIdentifier;
+import org.gradle.api.artifacts.ProjectDependency;
+import org.gradle.api.artifacts.dsl.DependencyHandler;
+import org.gradle.api.artifacts.result.ResolvedComponentResult;
 import org.gradle.api.tasks.Nested;
 
 public abstract class GetVersionPlugin implements Plugin<Project> {
 
-    private static final String GET_VERSION_ELEMENTS_CONFIGURATION_NAME = "gcvGetVersionElements";
-
-    static final String GET_VERSIONS_CAPABILITY = "gcv:get-versions:0";
+    private static final String GET_VERSIONS_CONFIGURATION_NAME = "gcvGetVersions";
 
     @Nested
     protected abstract GcvAttributes getAttributes();
 
     @Inject
+    protected abstract DependencyHandler getDependencies();
+
+    @Inject
     protected abstract ConfigurationContainer getConfigurations();
 
     @Override
-    public final void apply(Project rootProject) {
-        if (!rootProject.getRootProject().equals(rootProject)) {
-            throw new GradleException("GetVersionPlugin must be applied only to root project");
+    public final void apply(Project project) {
+        NamedDomainObjectProvider<Configuration> getVersions = getConfigurations()
+                .register(GET_VERSIONS_CONFIGURATION_NAME, configuration -> {
+                    configuration.setCanBeConsumed(false);
+                    configuration.setCanBeResolved(true);
+                    configuration.attributes(getAttributes()::configureGcvSourceAttributes);
+
+                    ProjectDependency rootDependency = (ProjectDependency) getDependencies()
+                            .project(Map.of("path", project.getRootProject().getPath()));
+                    rootDependency.capabilities(
+                            handler -> handler.requireCapabilities(VersionsLockPlugin.UNIFIED_CLASSPATH_CAPABILITY));
+                    configuration.getDependencies().add(rootDependency);
+                });
+
+        project.getExtensions().getExtraProperties().set("getVersion", new Closure<String>(project, project) {
+            /**
+             * Groovy will invoke this method if they just supply one arg, e.g. 'com.google.guava:guava'. This is the
+             * preferred signature because it's shortest.
+             */
+            @SuppressWarnings("for-rollout:UnusedMethod")
+            public String doCall(Object moduleVersion) {
+                return doCall(moduleVersion, getVersions.get());
+            }
+
+            /** Find a version from another configuration, e.g. from the gradle-docker plugin. */
+            @SuppressWarnings("for-rollout:FloggerArgumentToString")
+            public String doCall(Object moduleVersion, Configuration configuration) {
+                List<String> strings = Splitter.on(':').splitToList(moduleVersion.toString());
+                Preconditions.checkState(
+                        strings.size() == 2, "Expected 'group:name', found: %s", moduleVersion.toString());
+
+                return getVersion(project, strings.get(0), strings.get(1), configuration);
+            }
+
+            /** This matches the signature of nebula's dependencyRecommendations.getRecommendedVersion. */
+            @SuppressWarnings("for-rollout:UnusedMethod")
+            public String doCall(String group, String name) {
+                return getVersion(project, group, name, getVersions.get());
+            }
+
+            @SuppressWarnings("for-rollout:UnusedMethod")
+            public String doCall(String group, String name, Configuration configuration) {
+                return getVersion(project, group, name, configuration);
+            }
+        });
+    }
+
+    private static String getVersion(Project project, String group, String name, Configuration configuration) {
+        checkConfigurationBelongsToProject(project, configuration);
+        return getOptionalVersion(project, group, name, configuration)
+                .orElseThrow(() -> notFound(group, name, configuration));
+    }
+
+    private static void checkConfigurationBelongsToProject(Project project, Configuration configuration) {
+        Configuration ownConfiguration = project.getConfigurations().findByName(configuration.getName());
+        if (ownConfiguration != configuration) {
+            throw new GradleException(
+                    String.format("""
+                        Cannot call getVersion with %s as it does not belong to project '%s'. getVersion can only \
+                        resolve a configuration that lives in the project it is called from. Either supply a \
+                        configuration from '%s', or call getVersion from the project that owns %s.\
+                        """, configuration, project.getPath(), project.getPath(), configuration));
+        }
+    }
+
+    static Optional<String> getOptionalVersion(
+            Project project, String group, String name, Configuration configuration) {
+        if (GradleWorkarounds.isConfiguring(project.getState())) {
+            throw new GradleException(String.format("""
+                Not allowed to call gradle-consistent-versions's getVersion("%s", "%s", configurations.%s) at \
+                configuration time\
+                """, group, name, configuration.getName()));
         }
 
-        getConfigurations().register(GET_VERSION_ELEMENTS_CONFIGURATION_NAME, configuration -> {
-            configuration.setCanBeConsumed(true);
-            configuration.setCanBeResolved(false);
-            configuration.extendsFrom(getConfigurations()
-                    .getByName(VersionsLockPlugin.UNIFIED_CLASSPATH_DEPENDENCIES_CONFIGURATION_NAME));
-            configuration.getOutgoing().capability(GET_VERSIONS_CAPABILITY);
-            configuration.attributes(getAttributes()::configureGcvBaseAttributes);
-        });
+        List<ModuleVersionIdentifier> list =
+                configuration.getIncoming().getResolutionResult().getAllComponents().stream()
+                        .map(ResolvedComponentResult::getModuleVersion)
+                        .filter(item ->
+                                item.getGroup().equals(group) && item.getName().equals(name))
+                        .collect(toList());
 
-        rootProject
-                .getAllprojects()
-                .forEach(project -> project.getPluginManager().apply(GetVersionProjectPlugin.class));
+        if (list.isEmpty()) {
+            return Optional.empty();
+        }
+
+        if (list.size() > 1) {
+            throw new GradleException(
+                    String.format("Multiple modules matching '%s:%s' in %s: %s", group, name, configuration, list));
+        }
+
+        return Optional.of(Iterables.getOnlyElement(list).getVersion());
+    }
+
+    private static GradleException notFound(String group, String name, Configuration configuration) {
+        String actual = configuration.getIncoming().getResolutionResult().getAllComponents().stream()
+                .map(ResolvedComponentResult::getModuleVersion)
+                .map(mvi -> String.format("\t- %s:%s:%s", mvi.getGroup(), mvi.getName(), mvi.getVersion()))
+                .collect(Collectors.joining("\n"));
+        return new GradleException(String.format("""
+            Unable to find '%s:%s' in %s. This may happen if you specify the version in versions.props but do not have \
+            a dependency in the configuration. The configuration contained:
+            %s\
+            """, group, name, configuration, actual));
     }
 }
